requirements.txt
mysql-connector-python==8.0.33
python-dotenv==1.0.0
tabulate==0.9.0
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Werkzeug==3.0.1 